+++ src/Fable.Transforms/Python/Fable2Python.Transforms.fs
@@ -2659,6 +2659,9 @@ let rec transformAsExpr (com: IPythonCompiler) ctx (expr: Fable.Expr) : Expressi
         | Fable.Curry(e, arity) -> transformCurry com ctx e arity
         | Fable.Throw _
         | Fable.Debugger -> iife com ctx expr
+    | Fable.Quote(body, _isTyped, _r) ->
+        let emitted = QuotationEmitter.emitQuotedExpr com body
+        transformAsExpr com ctx emitted
 
 let transformAsSlice (com: IPythonCompiler) ctx expr (info: Fable.CallInfo) : Expression * Statement list =
     Expression.withStmts {
@@ -2980,6 +2983,10 @@ let rec transformAsStatements (com: IPythonCompiler) ctx returnStrategy (expr: F
         let target = com.GetIdentifierAsExpr(ctx, var.Name)
 
         [ Statement.for' (target = target, iter = iter, body = body) ]
+
+    | Fable.Quote(body, _isTyped, _r) ->
+        let emitted = QuotationEmitter.emitQuotedExpr com body
+        transformAsStatements com ctx returnStrategy emitted
 
 let transformFunction
     com